+++ src/main/java/com/mitarcon/connections/service/GreetingService.java
@@ -1,6 +1,5 @@
 package com.mitarcon.connections.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.mitarcon.connections.service.dto.GreetingDTO;
@@ -8,8 +7,8 @@ import com.mitarcon.connections.service.dto.GreetingDTO;
 public interface GreetingService {
 
 	Collection<GreetingDTO> findAll();
-	GreetingDTO FindOne(BigInteger id);
+	GreetingDTO FindOne(int id);
 	GreetingDTO create(GreetingDTO greeting);
 	GreetingDTO update(GreetingDTO greeting);
-	void delete(BigInteger id);
+	void delete(int id);
 }
